requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-mdeditor==0.1.20
djangorestframework==3.15.1
inflection==0.5.1
Markdown==3.6
mysqlclient==2.2.4
pillow==10.3.0
Pygments==2.17.2
pymemcache==4.0.0
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.1
